internal: adapt diagnostics to the new rename API
